updated the code for offline onboarding
