+++ apps/front-end/package.json
@@ -9,7 +9,8 @@
     "@fullcalendar/interaction": "^6.1.8",
     "@fullcalendar/react": "^6.1.8",
     "@fullcalendar/timegrid": "^6.1.8",
-    "@rjsf/core": "^5.13.0",  
+    "@reduxjs/toolkit": "^2.0.1",
+    "@rjsf/core": "^5.13.0",
     "@rjsf/utils": "^5.13.0",
     "@rjsf/validator-ajv8": "^5.13.0",
     "@shiksha/common-lib": "^1.0.0",
@@ -38,6 +39,7 @@
     "react-native-safe-area-context": "^3.4.1",
     "react-native-web": "^0.17.7",
     "react-qr-reader": "3.0.0-beta-1",
+    "react-redux": "^9.1.0",
     "react-router-dom": "^6.15.0",
     "react-scripts": "5.0.0",
     "react-zoom-pan-pinch": "^3.1.0",
