+++ apps/front-end/public/locales/en/translation.json
@@ -1261,6 +1261,7 @@
   "TRY_AADHAR_QR_KYC": "Try Aadhaar QR KYC",
   "TRY_AADHAR_UPLOAD_KYC": "Try Aadhaar upload KYC",
   "TRY_OTHER": "Try Other",
+  "TRY_AGAIN": "Try Again",
   "TYPE_OF_CAMPSITE ": "Type of Campsite ",
   "TYPE_OF_DEVICE": "Type of Device",
   "TYPE_OF_DOCUMENT": "Type of document",
