+++ apps/front-end/public/locales/hi/translation.json
@@ -1261,6 +1261,7 @@
   "TRY_AADHAR_QR_KYC": "आधार QR OKYC प्रयास करें",
   "TRY_AADHAR_UPLOAD_KYC": "आधार फोटो OKYC प्रयास करें",
   "TRY_OTHER": "कोई दूसरा कोशिश",
+  "TRY_AGAIN": "फिर से कोशिश करें",
   "TYPE_OF_CAMPSITE ": "शिविर स्थल का प्रकार ",
   "TYPE_OF_DEVICE": "डिवाइस का प्रकार",
   "TYPE_OF_DOCUMENT": "दस्तावेज़ का प्रकार",
