+++ apps/front-end/src/pages/front-end/FacilitatorOffline/FacilitatorOnboarding/FacilitatorOnboarding.jsx
@@ -1,5 +1,5 @@
-import { CustomOTPBox, FrontEndTypo, Layout, t } from "@shiksha/common-lib";
-import { Box, HStack, Image, Stack, VStack } from "native-base";
+import { FrontEndTypo, Layout, t } from "@shiksha/common-lib";
+import { Stack, VStack } from "native-base";
 import React, { useState, useCallback, useEffect } from "react";
 import validator from "@rjsf/validator-ajv8";
 import { get, set } from "idb-keyval";
@@ -17,12 +17,26 @@ import Form from "@rjsf/core";
 import Steper from "component/Steper";
 import { useNavigate } from "react-router-dom";
 
+import * as formSchemas from "./onboarding.schema";
+
+import { debounce } from "lodash";
+
+const {
+  dateOfBirthSchema,
+  contactDetailsSchema,
+  basicDetailsSchema,
+  volunteerExperienceSchema,
+  jobExperienceSchema,
+  educationDetailsSchema,
+  otherDetailsSchema,
+  profilePictureSchema,
+} = formSchemas;
+
 const FacilitatorOnboarding = () => {
-  const [activeScreenName, setActiveScreenName] = useState();
-  const [mobileNumber, setMobileNumber] = useState("");
-  const [previousScreen, setPreviousScreen] = useState(0);
-  const [page, setPage] = React.useState(0);
-  const [yearsRange, setYearsRange] = React.useState([1980, 2030]);
+  const [activeScreenName, setActiveScreenName] = useState(null);
+  const [mobileNumber, setMobileNumber] = useState(null);
+  const [page, setPage] = useState(0);
+  const [yearsRange, setYearsRange] = useState([1980, 2030]);
 
   const [date_of_birth, set_date_of_birth] = useState({});
 
@@ -37,34 +51,58 @@ const FacilitatorOnboarding = () => {
 
   const [user_data, setUserData] = useState(null);
 
-  // useEffect(async () => {
-  //   if (user_data !== null) {
-  //     set_date_of_birth(user_data?.users?.dob);
-  //   }
-  //   set_date_of_birth((prevData) => ({
-  //     ...prevData,
-  //     date_of_birth: date_of_birth,
-  //   }));
-  // }, [user_data]);
-
   const [countLoad, setCountLoad] = useState(0);
 
-  const fetchData = useCallback(async () => {
-    if (countLoad === 0) {
-      setCountLoad(1);
-    } else if (countLoad === 1) {
-      let userData = await get("user_data");
-      setUserData(userData);
-      setCountLoad(2);
-    } else if (countLoad === 2) {
-      setCountLoad(3);
-    }
-  }, [countLoad]);
-
   useEffect(() => {
-    fetchData();
-  }, [fetchData]);
+    const fetchData = async () => {
+      try {
+        console.log("Fetching data from IndexedDB...");
+        let userData = await get("user_data");
+        console.log("Fetched data:", userData);
 
+        // Update form data states here
+        setUserData(userData);
+
+        // Example of updating date of birth state
+        if (userData?.users?.dob) {
+          set_date_of_birth({
+            dob: userData.users.dob,
+            mobile: userData.users.mobile,
+            alternative_mobile_number: userData.users.alternative_mobile_number,
+            // Add other properties as needed
+          });
+        }
+
+        // if (userData?.users?.mobile) {
+        //   setFormDataContact({
+        //     mobile: userData.users.mobile,
+        //   });
+        // }
+
+        if (user_data?.core_faciltator?.device_ownership) {
+          setFormDataContact({
+            device_ownership: userData.formDataContact.device_ownership,
+            device_type: userData.formDataContact.device_type,
+          });
+        }
+        if (userData?.users?.gender) {
+          // Example of updating basic details state
+          setFormDataBasicDetails({
+            gender: userData.users.gender,
+            marital_status: userData.extended_users?.marital_status || "", // Add other properties as needed
+          });
+        }
+
+        // Update other form data states in a similar way
+
+        setCountLoad(2);
+      } catch (error) {
+        console.error("Error fetching data from IndexedDB:", error);
+      }
+    };
+
+    fetchData();
+  }, []);
   const navigate = useNavigate();
 
   const handleInputChange = (value) => {
@@ -73,19 +111,10 @@ const FacilitatorOnboarding = () => {
 
   const handleNextScreen = (screenName) => {
     setActiveScreenName(screenName);
-
     setPage(page + 1);
-    console.log("page", page);
-
     setPage((prevPage) => {
       const nextPage = prevPage + 1;
-      // console.log("Current page:", nextPage);
-
-      // console.log("Navigating to:", screenName);
       navigate(`/offline/profile/${screenName}`);
-
-      // console.log("Next page:", nextPage);
-
       return nextPage;
     });
   };
@@ -112,7 +141,6 @@ const FacilitatorOnboarding = () => {
       console.log("previous screen");
     }
   };
-
   const handleDateOfBirth = async () => {
     try {
       const updatedUserData = {
@@ -123,10 +151,9 @@ const FacilitatorOnboarding = () => {
         },
       };
 
-      // Update the specific fields in IndexedDB
       await set("user_data", updatedUserData);
 
-      setUserData(updatedUserData); // Update the local state for immediate reflection
+      setUserData(updatedUserData);
       setPage((prevPage) => prevPage + 1);
       handleNextScreen("contactDetails");
     } catch (error) {
@@ -135,7 +162,6 @@ const FacilitatorOnboarding = () => {
   };
   const handleContactDetails = async () => {
     try {
-      // Update the specific fields in IndexedDB
       const updatedUserData = {
         ...user_data,
         users: {
@@ -143,7 +169,6 @@ const FacilitatorOnboarding = () => {
           mobile: formDataContact.mobile,
           alternative_mobile_number: formDataContact.alternative_mobile_number,
           email_id: formDataContact.email_id,
-          // Add other fields as needed
         },
         core_faciltator: {
           ...user_data?.core_faciltator,
@@ -151,11 +176,9 @@ const FacilitatorOnboarding = () => {
           device_type: formDataContact.device_type,
         },
       };
-
-      // Update the specific fields in IndexedDB
       await set("user_data", updatedUserData);
 
-      setUserData(updatedUserData); // Update the local state for immediate reflection
+      setUserData(updatedUserData);
       setPage((prevPage) => prevPage + 1);
       handleNextScreen("basicDetails");
     } catch (error) {
@@ -169,7 +192,6 @@ const FacilitatorOnboarding = () => {
         users: {
           ...user_data?.users,
           gender: formDataBasicDetails.gender,
-          // Add other fields as needed
         },
         extended_users: {
           ...user_data?.extended_users,
@@ -177,11 +199,8 @@ const FacilitatorOnboarding = () => {
           social_category: formDataBasicDetails.social_category,
         },
       };
-
-      // Update the specific fields in IndexedDB
       await set("user_data", updatedUserData);
-
-      setUserData(updatedUserData); // Update the local state for immediate reflection
+      setUserData(updatedUserData);
       setPage((prevPage) => prevPage + 1);
       handleNextScreen("voluenteerExperience");
     } catch (error) {
@@ -189,40 +208,44 @@ const FacilitatorOnboarding = () => {
     }
   };
 
-  const handleVoluenteerExperience = async () => {
-    try {
-      const updatedUserData = {
-        ...user_data,
-        experience: [
-          {
-            ...user_data?.experience,
-            type: formDataVolunteerExperience.type,
-            role_title: formDataVolunteerExperience.role_title,
-            organization: formDataVolunteerExperience.organization,
-            experience_in_years:
-              formDataVolunteerExperience.experience_in_years,
-            related_to_teaching:
-              formDataVolunteerExperience.related_to_teaching,
-            description: formDataVolunteerExperience.description,
-          },
-        ],
-      };
+  const handleVolunteerExperience = async () => {
+    const propertyNames = [
+      "role_title",
+      "organization",
+      "experience_in_years",
+      "related_to_teaching",
+      "description",
+    ];
 
-      // Update the specific fields in IndexedDB
-      await set("user_data", updatedUserData);
+    const newExperiences =
+      formDataVolunteerExperience?.vo_experience?.vo_experience?.map(
+        (entry) => {
+          const newExperience = {};
+          for (const propName of propertyNames) {
+            newExperience[propName] = entry[propName] || "";
+          }
+          return newExperience;
+        }
+      ) || [];
 
-      setUserData(updatedUserData); // Update the local state for immediate reflection
-      setPage((prevPage) => prevPage + 1);
-      console.log("hi");
-      handleNextScreen("jobExperience");
-    } catch (error) {
-      console.error("Error storing data in IndexedDB:", error);
-    }
+    const experienceArray = Array.isArray(user_data.experience)
+      ? user_data.experience
+      : [];
+    console.log(experienceArray, "Experience");
+    const updatedUserData = {
+      ...user_data,
+      experience: [...experienceArray, ...newExperiences],
+    };
+    console.log(updatedUserData.experience, "After Update");
+
+    await set("user_data", updatedUserData);
+    setUserData(updatedUserData);
+    setPage((prevPage) => prevPage + 1);
+    console.log("hi");
+    handleNextScreen("jobExperience");
   };
 
-  //screen1
-
-  React.useEffect(() => {
+  useEffect(() => {
     let minYear = moment().subtract("years", 30);
     let maxYear = moment().subtract("years", 12);
     setYearsRange([minYear.year(), maxYear.year()]);
@@ -239,8 +262,8 @@ const FacilitatorOnboarding = () => {
     },
   };
 
-  const dateOfBirth = () => (
-    <>
+  const DateOfBirth = React.memo(() => {
+    return (
       <VStack flex={3} space={6}>
         <Stack space={3}>
           <Form
@@ -249,18 +272,7 @@ const FacilitatorOnboarding = () => {
             onSubmit={handleDateOfBirth}
             {...{ templates, FieldTemplate }}
             validator={validator}
-            schema={{
-              type: "object",
-              required: ["dob"],
-              properties: {
-                dob: {
-                  type: "string",
-                  label: "DATE_OF_BIRTH",
-                  description: "AS_PER_AADHAAR",
-                  format: "date",
-                },
-              },
-            }}
+            schema={dateOfBirthSchema}
             uiSchema={uiSchema}
           >
             <FrontEndTypo.Primarybutton
@@ -272,8 +284,8 @@ const FacilitatorOnboarding = () => {
           </Form>
         </Stack>
       </VStack>
-    </>
-  );
+    );
+  });
   const contactDetails = () => (
     <>
       <VStack flex={3} space={6}>
@@ -284,43 +296,7 @@ const FacilitatorOnboarding = () => {
           widgets={{ RadioBtn, CustomR }}
           {...{ templates, FieldTemplate, widgets }}
           validator={validator}
-          schema={{
-            type: "object",
-            properties: {
-              mobile: {
-                label: "CONTACT_INFORMATION",
-                type: "number",
-                title: "MOBILE_NUMBER",
-              },
-              device_ownership: {
-                label: "DEVICE_OWNERSHIP",
-                type: "string",
-                format: "RadioBtn",
-                enum: ["YES", "NO_I_USE_A_FAMILY_MEMBERS"],
-              },
-              device_type: {
-                label: "TYPE_OF_MOBILE_PHONE",
-                type: "string",
-                format: "CustomR",
-                grid: 2,
-                icons: [{ name: "AndroidLineIcon" }, { name: "AppleLineIcon" }],
-                enumNames: ["ANDROID", "IPHONE"],
-                enum: ["android", "iphone"],
-              },
-              alternative_mobile: {
-                label: "ALTERNATIVE_NUMBER",
-                type: "number",
-                title: "MOBILE_NUMBER",
-              },
-
-              email_id: {
-                label: "EMAIL_ID",
-                type: "string",
-                format: "email",
-                title: "EMAIL_ID",
-              },
-            },
-          }}
+          schema={contactDetailsSchema}
         >
           <VStack space={7}>
             <FrontEndTypo.Primarybutton
@@ -340,7 +316,6 @@ const FacilitatorOnboarding = () => {
       </VStack>
     </>
   );
-
   const basicDetails = () => (
     <>
       <VStack flex={3} space={6}>
@@ -351,59 +326,7 @@ const FacilitatorOnboarding = () => {
           widgets={{ RadioBtn, CustomR }}
           {...{ templates, FieldTemplate, widgets }}
           validator={validator}
-          schema={{
-            type: "object",
-            properties: {
-              gender: {
-                label: "GENDER",
-                type: "string",
-                format: "CustomR",
-                grid: 3,
-                icons: [
-                  {
-                    name: "Female",
-                    _icon: { size: "30" },
-                  },
-                  {
-                    name: "Male",
-                    _icon: { size: "30" },
-                  },
-                  {
-                    name: "Other",
-                    _icon: { size: "30" },
-                  },
-                ],
-                enumNames: ["FEMALE", "MALE", "OTHER"],
-                enum: ["female", "male", "other"],
-              },
-              marital_status: {
-                label: "MARITAL_STATUS",
-                type: "string",
-                format: "CustomR",
-                grid: 2,
-                enumNames: [
-                  "MARITAL_STATUS_MARRIED",
-                  "MARITAL_STATUS_UNMARRIED",
-                  "MARITAL_STATUS_SINGLE",
-                ],
-                enum: ["married", "unmarried", "single"],
-              },
-              social_category: {
-                label: "SOCIAL_CATEGORY",
-                type: "string",
-                format: "CustomR",
-                grid: 2,
-                enumNames: [
-                  "BENEFICIARY_SOCIAL_STATUS_GENERAL",
-                  "BENEFICIARY_SOCIAL_STATUS_SC",
-                  "BENEFICIARY_SOCIAL_STATUS_ST",
-                  "BENEFICIARY_SOCIAL_STATUS_OBC",
-                  "BENEFICIARY_REASONS_FOR_DROPOUT_REASONS_OTHER",
-                ],
-                enum: ["general", "sc", "st", "obc", "other"],
-              },
-            },
-          }}
+          schema={basicDetailsSchema}
         >
           <VStack space={4}>
             <FrontEndTypo.Primarybutton
@@ -432,75 +355,17 @@ const FacilitatorOnboarding = () => {
         <Form
           formData={formDataVolunteerExperience}
           onChange={(e) => setFormDataVolunteerExperience(e.formData)}
-          onSubmit={handleVoluenteerExperience}
+          onSubmit={handleVolunteerExperience}
           widgets={{ RadioBtn, CustomR }}
           {...{ templates, FieldTemplate, widgets }}
           validator={validator}
           uiSchema={uiSchema}
-          schema={{
-            type: "object",
-            required: ["aadharName"],
-            properties: {
-              volunteer_experience: {
-                type: "object",
-                title: "DO_YOU_HAVE_ANY_VOLUNTEER_EXPERIENCE",
-                properties: {
-                  vo_experience: {
-                    type: "array",
-                    items: {
-                      // title: "VOLUNTEER_EXPERIENCE",
-                      required: [
-                        "role_title",
-                        "organization",
-                        "experience_in_years",
-                      ],
-                      properties: {
-                        role_title: {
-                          title: "VOLUNTEER_TITLE",
-                          type: "string",
-                        },
-                        organization: {
-                          title: "COMPANY_AND_ORGANIZATION_NAME",
-                          type: "string",
-                        },
-                        add_description: {
-                          title: "THE_ENROLLMENT_DETAILS",
-                          type: "string",
-                        },
-
-                        experience_in_years: {
-                          label: "EXPERIENCE_IN_YEARS",
-                          type: "string",
-                          format: "CustomR",
-                          grid: 5,
-                          enumNames: ["<=1", "2", "3", "4", "5+"],
-                          enum: ["1", "2", "3", "4", "5"],
-                        },
-
-                        related_to_teaching: {
-                          label: "IS_THE_JOB_RELATED_TO_TEACHING",
-                          type: "string",
-                          format: "RadioBtn",
-                          enumNames: ["YES", "NO"],
-                          enum: ["yes", "no"],
-                        },
-                        description: {
-                          title: "DESCRIPTION",
-                          type: "string",
-                          // format: "Textarea",
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          }}
+          schema={volunteerExperienceSchema}
         >
           <VStack space={3}>
             <FrontEndTypo.Primarybutton
               style={{ background: "#FF0000", top: "50px" }}
-              onPress={handleVoluenteerExperience}
+              onPress={handleVolunteerExperience}
             >
               {t("SAVE_AND_NEXT")}
             </FrontEndTypo.Primarybutton>
@@ -528,60 +393,7 @@ const FacilitatorOnboarding = () => {
           {...{ templates, FieldTemplate, widgets }}
           validator={validator}
           uiSchema={uiSchema}
-          schema={{
-            type: "object",
-            required: ["aadharName"],
-            properties: {
-              job_experience: {
-                title: "DO_YOU_HAVE_ANY_JOB_EXPERIENCE",
-                type: "object",
-                properties: {
-                  experience: {
-                    type: "array",
-                    items: {
-                      title: "EXPERIENCE",
-                      required: [
-                        "role_title",
-                        "organization",
-                        "experience_in_years",
-                      ],
-                      properties: {
-                        role_title: {
-                          title: "JOB_TITLE",
-                          type: "string",
-                        },
-                        organization: {
-                          title: "COMPANY_NAME",
-                          type: "string",
-                        },
-                        description: {
-                          title: "DESCRIPTION",
-                          type: "string",
-                          format: "Textarea",
-                          rows: 5,
-                        },
-                        experience_in_years: {
-                          label: "EXPERIENCE_IN_YEARS",
-                          type: "string",
-                          format: "CustomR",
-                          grid: 5,
-                          enumNames: ["<=1", "2", "3", "4", "+5"],
-                          enum: ["1", "2", "3", "4", "5"],
-                        },
-                        related_to_teaching: {
-                          label: "IS_THE_JOB_RELATED_TO_TEACHING",
-                          type: "string",
-                          format: "RadioBtn",
-                          enumNames: ["YES", "NO"],
-                          enum: ["yes", "no"],
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          }}
+          schema={jobExperienceSchema}
         >
           <VStack space={3}>
             <FrontEndTypo.Primarybutton
@@ -610,44 +422,7 @@ const FacilitatorOnboarding = () => {
           {...{ templates, FieldTemplate, widgets }}
           validator={validator}
           uiSchema={uiSchema}
-          schema={{
-            title: "YOUR_HIGHEST_QUALIFICATION",
-            type: "object",
-            properties: {
-              qualification_master_id: {
-                type: ["string", "number"],
-                format: "CustomR",
-                grid: 2,
-                enumNames: [
-                  "PURSUING_GRADUATION",
-                  "GRADUATION",
-                  "MASTERS",
-                  "PHD",
-                  "12TH_GRADE",
-                ],
-                enum: [
-                  "persuing_graduation",
-                  "graduation",
-                  "masters",
-                  "phd",
-                  "12th_grade",
-                ],
-              },
-              degree: {
-                label: "DO_YOU_HAVE_ANY_TEACHING_DEGREE",
-                type: ["string", "number"],
-                format: "CustomR",
-                grid: 2,
-                enumNames: [
-                  "NTT",
-                  "D.El.Ed",
-                  "DROPOUT_REASONS_OTHER",
-                  "NO_ONE",
-                ],
-                enum: ["ntt", "ded", "other", "no"],
-              },
-            },
-          }}
+          schema={educationDetailsSchema}
         >
           <VStack space={3}>
             <FrontEndTypo.Primarybutton
@@ -676,33 +451,7 @@ const FacilitatorOnboarding = () => {
           {...{ templates, FieldTemplate, widgets }}
           validator={validator}
           uiSchema={uiSchema}
-          schema={{
-            title: "YOUR_WORK_AVAILABILITY_WILL_BE",
-            type: "object",
-            properties: {
-              qualification_master_id: {
-                type: ["string", "number"],
-                format: "CustomR",
-                grid: 2,
-                enumNames: ["FACILITATOR_PART_TIME", "FACILITATOR_FULL_TIME"],
-                enum: ["part_time", "full_time"],
-              },
-              degree: {
-                label: "HOW_DID_YOU_FIND_OUT_ABOUT_PROJECT_PRAGATI",
-                type: ["string", "number"],
-                format: "CustomR",
-                grid: 2,
-                enumNames: [
-                  "अन्य प्रेरक से",
-                  "NGO_REFERRAL",
-                  "विज्ञापन",
-                  "मैं स्वयं एक पुराना प्रेरक हूं",
-                  "DROPOUT_REASONS_OTHER",
-                ],
-                enum: ["ntt", "ded", "other", "no", ""],
-              },
-            },
-          }}
+          schema={otherDetailsSchema}
         >
           <VStack space={3}>
             <FrontEndTypo.Primarybutton
@@ -731,19 +480,7 @@ const FacilitatorOnboarding = () => {
           {...{ templates, FieldTemplate, widgets }}
           validator={validator}
           uiSchema={uiSchema}
-          schema={{
-            title: "YOUR_WORK_AVAILABILITY_WILL_BE",
-            type: "object",
-            properties: {
-              qualification_master_id: {
-                type: ["string", "number"],
-                format: "CustomR",
-                grid: 2,
-                enumNames: ["FACILITATOR_PART_TIME", "FACILITATOR_FULL_TIME"],
-                enum: ["part_time", "full_time"],
-              },
-            },
-          }}
+          schema={profilePictureSchema}
         >
           <VStack space={3}>
             <FrontEndTypo.Secondarybutton
@@ -760,7 +497,7 @@ const FacilitatorOnboarding = () => {
   const renderSwitchCase = () => {
     switch (activeScreenName) {
       case "dateOfBirth":
-        return dateOfBirth();
+        return <DateOfBirth />;
       case "contactDetails":
         return contactDetails();
       case "basicDetails":
@@ -776,7 +513,7 @@ const FacilitatorOnboarding = () => {
       case "facilitatorProfilePicture":
         return facilitatorProfilePicture();
       default:
-        return dateOfBirth();
+        return <DateOfBirth />;
     }
   };
