+++ apps/front-end/src/pages/front-end/FacilitatorOffline/FacilitatorRegistration/schema.js
@@ -1,70 +1,42 @@
-export default {
-  description: "1.Basic details",
-  type: "step",
+export const idVerificationSchema = {
+  type: "object",
+  required: ["aadharName"],
   properties: {
-    1: {
-      title: "CONTACT_DETAILS",
+    aadharName: {
+      type: "string",
+      title: "AADHAR_CARD",
+    },
+  },
+};
 
-      type: "object",
-      required: ["mobile"],
-      properties: {
-        mobile: {
-          type: "number",
-          title: "MOBILE_NUMBER",
-        },
-        mark_as_whatsapp_number: {
-          type: "string",
-          label: "MARK_AS_WHATSAPP_NO",
-          format: "RadioBtn",
+export const enterBasicDetailsSchema = {
+  type: "object",
+  required: ["firstName", "lastName"],
 
-          enumNames: ["YES", "NO"],
-          enum: ["yes", "no"],
-        },
-        device_type: {
-          type: "string",
-          label: "TYPE_OF_MOBILE_PHONE",
-          format: "CustomR",
-          enumNames: ["SMARTPHONE", "BASIC"],
-          enum: ["smartphone", "basic"],
-        },
-        device_ownership: {
-          type: "string",
-          label: "MARK_OWNERSHIP",
-          format: "RadioBtn",
-          enumNames: ["SELF", "FAMILY_MEMBER", "NEIGHBOUR", "OTHER"],
-          enum: ["self", "family_member", "neighbour", "other"],
-        },
+  properties: {
+    firstName: {
+      type: "string",
+      title: "FIRST_NAME",
+    },
+    lastName: {
+      type: "string",
+      title: "LAST_NAME",
+    },
+    mobile: {
+      label: "HOW_CAN_CONTACT_YOU",
+      type: "number",
+      title: "MOBILE_NUMBER",
+    },
+  },
+};
 
-        alternative_mobile_number: {
-          type: ["number", "null"],
-          title: "ALTERNATIVE_NUMBER",
-        },
-        alternative_device_type: {
-          label: "TYPE_OF_MOBILE_PHONE",
-          format: "CustomR",
-          type: "string",
-          enumNames: ["SMARTPHONE", "BASIC"],
-          enum: ["smartphone", "basic"],
-        },
-        alternative_device_ownership: {
-          type: "string",
-          label: "MARK_OWNERSHIP",
-          format: "RadioBtn",
-          enumNames: ["SELF", "FAMILY_MEMBER", "NEIGHBOUR", "OTHER"],
-          enum: ["self", "family_member", "neighbour", "other"],
-        },
-
-        email_id: {
-          type: "string",
-          format: "email",
-          title: "EMAIL_ID",
-        },
-        edit_page_type: {
-          type: "string",
-          default: "edit_contact",
-          format: "hidden",
-        },
-      },
+export const contactDetailsSchema = {
+  type: "object",
+  required: ["aadharName"],
+  properties: {
+    mobile: {
+      type: "number",
+      title: "MOBILE_NUMBER",
     },
   },
 };
