+++ apps/front-end/src/pages/front-end/FacilitatorOffline/FacilitatorRegistration/styles.js
@@ -1,0 +1,59 @@
+const stylesheet = {
+  text1: {
+    fontFamily: "Inter",
+    fontWeight: "400",
+    fontSize: "14px",
+    lineHeight: "26px",
+    color: "#3F8BF1",
+    textDecoration: "underline",
+    marginTop: "5px",
+  },
+  image: {},
+  buttonStyle: {
+    height: "91px",
+    padding: "16px",
+  },
+  headerText: {
+    color: "var(--Gray-90, #212121);",
+    fontFamily: "Inter",
+    fontWeight: "700",
+    fontSize: "22px",
+  },
+  languageButton: {
+    width: 150,
+    borderRadius: 14,
+    backgroundColor: "var(--Gray-10, #FAFAFA)",
+    borderWidth: 2,
+    borderColor: "var(--Secondary-Blue, #084B82)",
+    height: 91,
+  },
+  languageButtonText: {
+    color: "var(--Secondary-Blue, #084B82)",
+    textAlign: "center",
+    fontFamily: "Inter",
+    fontSize: 12,
+    fontStyle: "normal",
+    fontWeight: "600",
+    lineHeight: "normal",
+  },
+
+  defaultLanguageButton: {
+    width: 150,
+    borderRadius: 4,
+    backgroundColor: "var(--Gray-10, #FAFAFA)",
+    border: " 1px solid var(--Gray-30, #E0E0E0)",
+    borderColor: "gray",
+    height: 91,
+  },
+
+  defaulutLanguageButtonText: {
+    color: "var(--Gray-80, #424242)",
+    textAlign: "center",
+    fontFamily: "Inter",
+    fontSize: 12,
+    fontStyle: "normal",
+    fontWeight: "500",
+    lineHeight: "normal",
+  },
+};
+export default stylesheet;
